backend/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
argcomplete==3.6.3
async-timeout==5.0.1
attrs==25.4.0
bcrypt==5.0.0
beanie==1.30.0
cachecontrol==0.14.4
certifi==2026.1.4
cffi==2.0.0
click==8.3.1
colorama==0.4.6
coverage==7.13.2
cryptography==46.0.5
dnspython==2.8.0
ecdsa==0.19.1
email-validator==2.3.0
fastapi==0.128.0
fastapi-cli==0.0.20
fastapi-cloud-cli==0.11.0
fastar==0.8.0
frozenlist==1.7.0
h11==0.16.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
idna==3.11
iniconfig==2.3.0
jaraco-classes==3.4.0
Jinja2==3.1.6
lazy-model==0.2.0
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
motor==3.7.1
msgpack==1.1.2
mslex==1.3.0
multidict==6.7.0
packaging==26.0
pipx==1.8.0
platformdirs==4.5.1
pluggy==1.6.0
propcache==0.3.1
psutil==6.1.1
ptyprocess==0.7.0
pyasn1==0.6.2
pycparser==2.22
pydantic==2.12.5
pydantic-extra-types==2.11.0
pydantic-settings==2.12.0
pydantic_core==2.41.5
Pygments==2.19.2
pymongo==4.16.0
pytest==9.0.2
pytest-cov==7.0.0
python-dotenv==1.2.1
python-jose==3.5.0
python-multipart==0.0.22
PyYAML==6.0.3
rich==14.3.1
rich-toolkit==0.17.2
rignore==0.7.6
rsa==4.9.1
ruff==0.14.14
sentry-sdk==2.51.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
starlette==0.50.0
taskipy==1.14.1
tomli==2.4.0
tornado==6.5.4
typer==0.21.1
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
userpath==1.9.2
uvicorn==0.40.0
uvloop==0.22.1
watchfiles==1.1.1
websockets==16.0
yarl==1.22.0
pytest_asyncio==1.3.0
langchain==1.2.10
langchain-ollama==1.0.1